+++ c4data/bb7Data/BB7FebexData.h
@@ -10,27 +10,27 @@ class BB7FebexItem : public TObject
 
         void SetAll(int64_t wr,
                     uint16_t id,
-                    uint8_t b_id,
+                    uint32_t b_id,
                     uint64_t ev_time,
-                    uint8_t ch_id,
+                    uint32_t ch_id,
                     uint64_t ch_time,
                     bool p,
                     bool o,
                     uint32_t ch_energy,
-                    uint8_t ch_id_traces,
+                    uint32_t ch_id_traces,
                     std::vector<uint16_t> tr);
         void Reset();
 
         int64_t Get_wr_t() const;
         uint16_t Get_wr_id() const;
-        uint8_t Get_board_id() const;
+        uint32_t Get_board_id() const;
         uint64_t Get_board_event_time() const;
-        uint8_t Get_channel_id() const;
+        uint32_t Get_channel_id() const;
         uint64_t Get_channel_time() const;
         bool Get_pileup() const;
         bool Get_overflow() const;
         uint32_t Get_channel_energy() const;
-        uint8_t Get_channel_id_traces() const;
+        uint32_t Get_channel_id_traces() const;
         std::vector<uint16_t> Get_trace() const;
 
         // Getters
@@ -39,14 +39,14 @@ class BB7FebexItem : public TObject
     
         int64_t wr_t;
         uint16_t wr_id;
-        uint8_t board_id;
+        uint32_t board_id;
         uint64_t board_event_time;
-        uint8_t channel_id;
+        uint32_t channel_id;
         uint64_t channel_time;
         bool pileup;
         bool overflow;
         uint32_t channel_energy;
-        uint8_t channel_id_traces;
+        uint32_t channel_id_traces;
         std::vector<uint16_t> trace; 
 
 };
@@ -61,7 +61,7 @@ inline uint16_t BB7FebexItem::Get_wr_id() const
     return wr_id;
 }
 
-inline uint8_t BB7FebexItem::Get_board_id() const
+inline uint32_t BB7FebexItem::Get_board_id() const
 {
     return board_id;
 }
@@ -71,7 +71,7 @@ inline uint64_t BB7FebexItem::Get_board_event_time() const
     return board_event_time;
 }
 
-inline uint8_t BB7FebexItem::Get_channel_id() const
+inline uint32_t BB7FebexItem::Get_channel_id() const
 {
     return channel_id;
 }
@@ -96,7 +96,7 @@ inline uint32_t BB7FebexItem::Get_channel_energy() const
     return channel_energy;
 }
 
-inline uint8_t BB7FebexItem::Get_channel_id_traces() const
+inline uint32_t BB7FebexItem::Get_channel_id_traces() const
 {
     return channel_id_traces;
 }
